src/tasks/dbbench/requirements.txt
mysql-connector-python==8.0.33
docker==6.0.1